+++ src/key.ts
@@ -39,7 +39,7 @@ export class Key {
   _isLedger: boolean;
   _ledgerPath: string;
   _ledgerCurve: number;
-  ready: Promise<void>;
+  ready: Promise<boolean>;
 
   constructor({
     key,